+++ src/Frontend/Dfe.Complete/Constants/RouteConstants.cs
@@ -14,7 +14,9 @@
 
         public const string ProjectsByRegion = "/projects/all/regions/{0}";
 
+        // Your team projects
         public const string TeamProjectsInProgress = "/projects/team/in-progress";
+        public const string TeamProjectsCompleted = "/projects/team/completed";
         public const string TeamProjectsUnassigned = "/projects/team/unassigned";
 
         // Project